+++ doc/source/doxygen-docs/changelog.md
@@ -12,6 +12,8 @@
     - New mrpt::obs::CActionCollection::insert() overload for smart pointers.
     - New method mrpt::obs::CObservation2DRangeScan::getScanAngle() and clarify docs on class members.
     - New class mrpt::obs::CObservation3DScene.
+  - \ref mrpt_opengl_grp
+    - Deprecate mrpt::opengl::COpenGLScene::dumpListOfObjects() in favor of new mrpt::opengl::COpenGLScene::asYAML()
 - BUG FIXES:
   - Fix wrong formatting of empty *string* values (not *null* values) in mrpt::containers::yaml.